Ensure we use application/octet-stream for file upload.
Nexus, at the minimum, will fail to process the uploaded file if we
omit this header (I assume it considers the incoming body as text vs
a stream of bytes).

Nexus also doesn't return a body on succesful create. This caused
the upload_result to be empty, which made the artifact value empty,
which caused a failure. Always create a valid return value
that describes the HTTP status code, the reason, and any body.